Initial commit for composition work
This commit shows the diff necessary to fork printSchema to achieve
parity with our existing printFederatedSchema implementation.

The purpose of this first step is to update our fork using the most
current version of printSchema, since it's changed over time.

In future commits, this new function will diverge from the existing
implementation in order to implement a new composition format.

Note: newly introduced test cases are in place to validate work in an
iterative manner and are subject to updates and/or deletion.