Port updated fork with parity into existing implementation
This updated fork resolves a long outstanding issue in the existing
fork where block string comments (""") do not allow for double
quotes to exist inside of them. This has been a pain point for a
number of our users, forcing them to manually or scriptually parse
their schema for these occurences and remove them in order to register
their services.

Fixes #4339
Fixes #3864